Move webllm hooks to utils under hooks folder from components hooks
